Add Stalwart 0.16 JMAP admin compatibility (#628, #1168) (#1131)
* Split up MailClient from clients.py into 3 different MailClient (incl interface.)

* Some initial jmap work

Not happy with the current state, but also not happy with existing jmap
clients. The idea is to make this pretty lean but still allow for some
developer ux wins.

* Kick-off stalwart types, and start with get domain

* WIP integration of pydantic

* Fix jmap calls with, and add get_account

* Add a few more types, wrangle `@type`, and add a few more methods to stalwart client jmap.

* Base types off of BaseSchema and adjust properties to snake_case because python

* Finish up create_account

* Refactor clients test structure

* Fix create_account's return, quota param, and deprecate app password. (It's not used.)

* Update docblock for mail_client_jmap.py

* Cache primary domain id, and fix get_account / add delete_account

* * Move types into `mail.types.*` instead of being *_types.py inside client.
* Rename AccountType/DomainType to just Account/Domain.

* Include test script

* * Add update_individual
* Port jmap types to pydantic

* Add save_app_password

* Refactor how the jmap classes are setup so we don't confused admin calls with user calls.

* Move domain lookup to its own function and replace the instance in create_account

* WIP refactor to split up basic jmap ops

* Refactor jmap clients, and fix up save_email_addresses

* Refactor invocation names to the constants

* Implement replace/delete email addresses, fix delete account, and swap exclude_none to exclude_unset

* Add StorageQuota type, and implement update_quota

* Refactor and add DkimSignature type

* Move JMapType and Id into jmap types, and add get_identity with types.

* Create private key and implement create_dkim for dkim1 signatures

* Remove key dumper test code

* Add delete_dkim

* Implement most of the dkim/dns functions, and modify some return types.

* Re-add function I accidentally cut

* Merge stalwart v0.16 into main docker file so I can start running tests.

Config is stored inside sqlite file, hopefully this is cross-compat...

* Swap listener to 8180 for http

* Add some minor hacks to get the dashboard working

* Get domain verification working-ish for migrated users

* Move main.db to main.db.example

* Move config.json to config.json.example

* Only volume mount the db file not the folder

Otherwise we'll run into "attempt to write a readonly database" error.

* Update bootstrap.py for v0.16

* Lint

* Format

* Fix import

* Foce test_legacy to use mailclient legacy

* `chmod 777 main.db` which gets around the cannot read readonly db error.

* Hook-up new stalwart url, user, pass, and auth method

* Add response dump files, an initial test and blank out debug dump

* Lint / Format

* Remove unused import

* Fix ruff format checks

---------

+++ src/thunderbird_accounts/mail/views.py
@@ -1,3 +1,4 @@
+from thunderbird_accounts.core.types import AuthenticatedHttpRequest
 import datetime
 import json
 import logging
@@ -43,6 +44,7 @@ from thunderbird_accounts.mail.models import Account, Email, Domain
 from thunderbird_accounts.mail import tasks as mail_tasks
 from thunderbird_accounts.mail import utils
 from thunderbird_accounts.subscription.decorators import active_subscription_required
+from thunderbird_accounts.mail.types import stalwart
 
 
 def _critical_errors_from_stale_dns_records(stale_dns_records: list[dict]) -> list[DomainVerificationErrors]:
@@ -140,7 +142,7 @@ def display_name_set(request: HttpRequest):
 @login_required
 @require_http_methods(['POST'])
 @active_subscription_required
-def create_custom_domain(request: HttpRequest):
+def create_custom_domain(request: AuthenticatedHttpRequest):
     """Creates a custom domain for the user"""
     data = json.loads(request.body)
     domain_name = data.get('domain-name')
@@ -168,6 +170,15 @@ def create_custom_domain(request: HttpRequest):
                 raise DomainAlreadyExistsError(domain_name)
         except DomainNotFoundError:
             pass
+
+        if request.user.is_migrated:
+            try:
+                # We need to create a disabled domain
+                stalwart_client.create_domain(domain_name, is_enabled=False)
+            except DomainAlreadyExistsError as ex:
+                raise ex
+
+        # FIXME: There may need to be clean up done if create_dkim raises an error.
 
         # If request fails the dkim will hit our general exception catch
         # we want this to happen before the local reference (Domain model) is created
@@ -266,7 +277,7 @@ def _domain_verification_error(domain, exc) -> JsonResponse:
 @login_required
 @require_http_methods(['POST'])
 @active_subscription_required
-def verify_custom_domain(request: HttpRequest):
+def verify_custom_domain(request: AuthenticatedHttpRequest):
     """Verifies a custom domain"""
     data = json.loads(request.body)
     domain_name = data.get('domain-name')
@@ -328,6 +339,9 @@ def verify_custom_domain(request: HttpRequest):
                 # Fetch or create a domain on Stalwart's end, and retrieve the domain_id
                 if stalwart_resp:
                     domain_id = stalwart_resp.get('id')
+                    # Now we need to enable the domain if it's not already enabled.
+                    if request.user.is_migrated and not stalwart_resp.is_enabled:
+                        stalwart_client.update_domain(domain_name, stalwart.DomainUpdate(is_enabled=True))
                 else:
                     domain_id = stalwart_client.create_domain(domain_name)
 
@@ -369,7 +383,7 @@ def verify_custom_domain(request: HttpRequest):
 @login_required
 @require_http_methods(['DELETE'])
 @active_subscription_required
-def remove_custom_domain(request: HttpRequest):
+def remove_custom_domain(request: AuthenticatedHttpRequest):
     """Removes a custom domain"""
     data = json.loads(request.body)
     domain_name = data.get('domain-name')
@@ -421,11 +435,12 @@ def remove_custom_domain(request: HttpRequest):
                     # so try deleting dkim and then local ref
                     _capture_domain_exception(ex, domain, phase='delete_stalwart_domain_not_found')
 
-            cleanup_phase = 'delete_dkim'
-            stalwart_client.delete_dkim(_domain.name)
+            if not request.user.is_migrated:
+                cleanup_phase = 'delete_dkim'
+                stalwart_client.delete_dkim(_domain.name)
 
-            cleanup_phase = 'delete_hosted_dkim_dns_records'
-            mail_tasks.delete_hosted_dkim_dns_records.delay(_domain.name)
+                cleanup_phase = 'delete_hosted_dkim_dns_records'
+                mail_tasks.delete_hosted_dkim_dns_records.delay(_domain.name)
             break
 
         cleanup_phase = 'delete_local_domain'
